requirements-dev.txt
Django~=4.0.1
django-allauth~=0.47.0
django-widgets-improved~=1.5.0
django-widget-tweaks~=1.4.12
Pillow~=8.4.0
django-boost~=1.7.2
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.3
django-heroku==0.3.1